Renamed telescope folder to telescopeIO
