+++ src/main/java/frc/robot/Robot.java
@@ -17,8 +17,8 @@ import frc.robot.subsystems.SwerveIO.module.SwerveModuleIOSim;
 import frc.robot.subsystems.SwerveIO.module.SwerveModuleIOSparkMAX;
 import frc.robot.subsystems.elevatorIO.Elevator;
 import frc.robot.subsystems.elevatorIO.ElevatorIOSim;
-import frc.robot.subsystems.telescope.Telescope;
-import frc.robot.subsystems.telescope.TelescopeIOSim;
+import frc.robot.subsystems.telescopeIO.Telescope;
+import frc.robot.subsystems.telescopeIO.TelescopeIOSim;
 import frc.robot.util.MotionHandler.MotionMode;
 import frc.robot.util.RedHawkUtil.ErrHandler;
 import org.littletonrobotics.junction.LoggedRobot;
