+++ src/main/java/frc/robot/subsystems/telescopeIO/TelescopeIO.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.telescope;
+package frc.robot.subsystems.telescopeIO;
 
 import org.littletonrobotics.junction.AutoLog;
