+++ src/main/java/frc/robot/subsystems/telescopeIO/TelescopeIOSparks.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.telescope;
+package frc.robot.subsystems.telescopeIO;
 
 import com.revrobotics.CANSparkMax;
